+++ elog.feishu.config.js
@@ -26,12 +26,27 @@ module.exports = {
   },
   image: {
     enable: true,
-    platform: 'local',
+    platform: process.env.IMAGE_PLATFORM || 'local',
     limit: 1,
     local: {
       outputDir: './src/feishu/images',
       prefixKey: 'images',
       pathFollowDoc: true,
+    },
+    github: {
+      token: process.env.GITHUB_TOKEN,
+      user: process.env.ELOG_GITHUB_USER,
+      repo: process.env.ELOG_GITHUB_REPO,
+      branch: 'main',
+      prefixKey: 'feishu',
+    },
+    cos: {
+      secretId: process.env.COS_SECRET_ID,
+      secretKey: process.env.COS_SECRET_KEY,
+      bucket: process.env.COS_BUCKET,
+      region: process.env.COS_REGION,
+      host: process.env.COS_HOST,
+      prefixKey: 'feishu',
     },
   }
 }